requirements.txt
boto3==1.40.32
python-dotenv==1.1.1
requests==2.32.5
pandas==2.3.2
